refactor: service 패키지를 domain으로 변경 및 접근 제어자 수정
- service 패키지명을 domain으로 변경하여 클래스들의 위치 명확화
- RacingGameController의 inputTryCount 메서드를 private으로 변경
