+++ src/main/kotlin/racingcar/domain/InputValidator.kt
@@ -1,4 +1,4 @@
-package racingcar.service
+package racingcar.domain
 
 import racingcar.resources.Messages
