+++ src/main/kotlin/racingcar/domain/RacingGameService.kt
@@ -1,4 +1,4 @@
-package racingcar.service
+package racingcar.domain
 
 import camp.nextstep.edu.missionutils.Randoms
 import racingcar.model.Car
